Make the form interactive
I've added functionality to make the feedback form interactive.

I moved the feedback form into its own component, `FeedbackForm`. Its responsibility is to render the feedback form and show/hide questions in response to the feedback given. When a user says the page wasn't useful, it asks further questions for more detailed feedback. Whereas if the page was useful, the form will automatically submit and no further questions are asked. Upon form submit, this component will emit a `submit` event.

The top-level `App` component listens for the `submit` event from `FeedbackForm`. Eventually it will save the submitted feedback to a database, however for now it simply emulates a save with a 500ms timeout. Once saved, a 'thank you' message is shown to the user.

I also made the static HTML markup more dynamic:
- The options under 'Why isn't this page useful?' are dynamic, using `v-for`
- `v-model` is used to capture the input values
- The submit button submits the form